requirements.txt
streamlit
google-generativeai
google-api-python-client
google-auth
requests
pandas
python-dotenv
PyYAML
